resolved review comments
